- Moved mixin accessor to mixin package
