+++ src/main/java/com/seacroak/basicweapons/mixin/PlayerEntityAccessor.java
@@ -1,4 +1,4 @@
-package com.seacroak.basicweapons;
+package com.seacroak.basicweapons.mixin;
 
 public interface PlayerEntityAccessor {
   float bw$getCooldown(float baseTime);
